operator/tls: extract code for kafka api tls
